New/KAVACH V1/requirements.txt
psutil==5.9.0
watchdog==2.1.6
requests==2.28.1
scapy==2.4.5
pandas==1.5.0
numpy==1.23.0
scikit-learn==1.1.2
joblib==1.2.0
dnspython==2.2.1